New chime6_audio_edit.json after bug fixes.
The new alignment is generated by a cleaner version of the MATLAB
alignment code and also
- fixes a bug that led to frame-drop location
estimates being slightly off (but not the frame-drop durations).
- fixes a bug the led to small error in the clock drift estimate
for device S18_06 and S24_06.

NOTE:
- This alignment still does not address the problems with
devices S01_U02 and S01_U05. The alignment for these devices
will still be up to 50 ms off. The fix requires a piecewise
linear fit and a new version of correct_signals_for_clock_drift.py.